+++ packages/node-package/src/index.js
@@ -1,6 +1,0 @@
-//--------------------------------------------------------
-//-- ESLint config
-//--------------------------------------------------------
-import loader from "@absolunet/eslint-loader";
-
-module.exports = loader(__dirname);