+++ packages/package/test/feature/index.test.js
@@ -1,9 +1,0 @@
-//--------------------------------------------------------
-//-- package - Feature tests
-//--------------------------------------------------------
-import configTests from "../../../../test/feature/config-tests";
-import { given, when, then } from "./index.gwt";
-
-describe(`Validate that 'package' works`, () => {
-	configTests({ given, when, then });
-});