+++ packages/package/test/generic/index.test.js
@@ -1,6 +1,0 @@
-//--------------------------------------------------------
-//-- Tester
-//--------------------------------------------------------
-import { tester } from "@absolunet/tester";
-
-tester.genericRepositoryTests();